# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cake Scheduler API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/cake/days": {
            "get": {
                "tags": [
                    "CakeEvent"
                ],
                "summary": "Get cake days",
                "operationId": "825d528117a3e873704f0d364c83f421",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "errorCode": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "cake_date": {
                                                        "type": "date-time"
                                                    },
                                                    "small_cakes": {
                                                        "type": "integer"
                                                    },
                                                    "large_cakes": {
                                                        "format": "integer"
                                                    },
                                                    "developers": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {}
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/birthday/upload": {
            "post": {
                "tags": [
                    "CakeEvent"
                ],
                "summary": "Upload developer's birthday list",
                "operationId": "546c915bd4f934a0af495f0c53ea8e71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "birthday_file": {
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "errorCode": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "CakeEvent",
            "description": "CakeEvent"
        }
    ]
}